+++ depthai_ros_driver/include/depthai_ros_driver/dai_nodes/sensors/rgbd.hpp
@@ -34,6 +34,7 @@ enum class RGBDLinkType { rgb, depth };
 };
 class SensorWrapper;
 class Stereo;
+class ToF;
 class RGBD : public BaseNode {
    public:
 
@@ -44,6 +45,13 @@ class RGBD : public BaseNode {
                      bool rsCompat,
                      SensorWrapper& camNode,
                      Stereo& stereoNode);
+    explicit RGBD(const std::string& daiNodeName,
+                     std::shared_ptr<rclcpp::Node> node,
+                     std::shared_ptr<dai::Pipeline> pipeline,
+                     std::shared_ptr<dai::Device> device,
+                     bool rsCompat,
+                     SensorWrapper& camNode,
+                     ToF& tofNode);
     ~RGBD();
     void updateParams(const std::vector<rclcpp::Parameter>& params) override;
     void setupQueues(std::shared_ptr<dai::Device> device) override;
