+++ depthai_ros_driver/src/dai_nodes/sensors/rgbd.cpp
@@ -1,14 +1,14 @@
 #include "depthai_ros_driver/dai_nodes/sensors/rgbd.hpp"
 
-#include <depthai/pipeline/node/Camera.hpp>
-#include <depthai/pipeline/node/ImageAlign.hpp>
-
 #include "depthai/device/Device.hpp"
 #include "depthai/pipeline/MessageQueue.hpp"
 #include "depthai/pipeline/Pipeline.hpp"
+#include "depthai/pipeline/node/Camera.hpp"
+#include "depthai/pipeline/node/ImageAlign.hpp"
 #include "depthai/pipeline/node/host/RGBD.hpp"
 #include "depthai_bridge/PointCloudConverter.hpp"
 #include "depthai_ros_driver/dai_nodes/sensors/stereo.hpp"
+#include "depthai_ros_driver/dai_nodes/sensors/tof.hpp"
 #include "depthai_ros_driver/param_handlers/rgbd_param_handler.hpp"
 #include "depthai_ros_driver/utils.hpp"
 #include "rclcpp/node.hpp"
@@ -37,7 +37,8 @@ RGBD::RGBD(const std::string& daiNodeName,
             std::pair<int, int>(ph->getOtherNodeParam<int>(camNode.getName(), "i_width"), ph->getOtherNodeParam<int>(camNode.getName(), "i_height")),
             dai::ImgFrame::Type::RGB888i,
             dai::ImgResizeMode::CROP,
-            30.0);
+            ph->getOtherNodeParam<float>(camNode.getName(), "i_fps"),
+            true);
         out->link(rgbdNode->inColor);
         align = pipeline->create<dai::node::ImageAlign>();
         stereo->depth.link(align->input);
@@ -48,11 +49,42 @@ RGBD::RGBD(const std::string& daiNodeName,
             std::pair<int, int>(ph->getOtherNodeParam<int>(camNode.getName(), "i_width"), ph->getOtherNodeParam<int>(camNode.getName(), "i_height")),
             dai::ImgFrame::Type::RGB888i,
             dai::ImgResizeMode::CROP,
-            30.0);
+            ph->getOtherNodeParam<float>(camNode.getName(), "i_fps"),
+            true);
         out->link(rgbdNode->inColor);
         out->link(stereo->inputAlignTo);
         stereo->depth.link(rgbdNode->inDepth);
     }
+
+    RCLCPP_DEBUG(getLogger(), "Node %s created", daiNodeName.c_str());
+}
+RGBD::RGBD(const std::string& daiNodeName,
+           std::shared_ptr<rclcpp::Node> node,
+           std::shared_ptr<dai::Pipeline> pipeline,
+           std::shared_ptr<dai::Device> device,
+           bool rsCompat,
+           SensorWrapper& camNode,
+           ToF& tofNode)
+    : BaseNode(daiNodeName, node, pipeline, device->getDeviceName(), rsCompat) {
+    RCLCPP_DEBUG(getLogger(), "Creating node %s", daiNodeName.c_str());
+    setNames();
+    rgbdNode = pipeline->create<dai::node::RGBD>()->build();
+    ph = std::make_unique<param_handlers::RGBDParamHandler>(node, daiNodeName, device->getDeviceName(), rsCompat);
+    ph->declareParams(rgbdNode);
+    auto color = camNode.getUnderlyingNode();
+    auto tof = tofNode.getUnderlyingNode();
+    auto platform = device->getPlatform();
+    auto* out = color->requestOutput(
+        std::pair<int, int>(ph->getOtherNodeParam<int>(camNode.getName(), "i_width"), ph->getOtherNodeParam<int>(camNode.getName(), "i_height")),
+        dai::ImgFrame::Type::RGB888i,
+        dai::ImgResizeMode::CROP,
+        ph->getOtherNodeParam<float>(camNode.getName(), "i_fps"),
+        true);
+    out->link(rgbdNode->inColor);
+    align = pipeline->create<dai::node::ImageAlign>();
+    tof->depth.link(align->input);
+    out->link(align->inputAlignTo);
+    align->outputAligned.link(rgbdNode->inDepth);
 
     RCLCPP_DEBUG(getLogger(), "Node %s created", daiNodeName.c_str());
 }
@@ -80,14 +112,14 @@ void RGBD::closeQueues() {
 }
 
 void RGBD::pclCB(const std::string& /*name*/, const std::shared_ptr<dai::ADatatype>& data) {
-        auto pclData = std::dynamic_pointer_cast<dai::PointCloudData>(data);
-        std::deque<sensor_msgs::msg::PointCloud2> deq;
-        pclConv->toRosMsg(pclData, deq);
-        while(deq.size() > 0) {
-            auto currMsg = deq.front();
-            pclPub->publish(currMsg);
-            deq.pop_front();
-        }
+    auto pclData = std::dynamic_pointer_cast<dai::PointCloudData>(data);
+    std::deque<sensor_msgs::msg::PointCloud2> deq;
+    pclConv->toRosMsg(pclData, deq);
+    while(deq.size() > 0) {
+        auto currMsg = deq.front();
+        pclPub->publish(currMsg);
+        deq.pop_front();
+    }
 }
 
 void RGBD::link(dai::Node::Input in, int /*linkType*/) {
