+++ depthai_ros_driver/src/param_handlers/tof_param_handler.cpp
@@ -14,9 +14,12 @@ ToFParamHandler::ToFParamHandler(std::shared_ptr<rclcpp::Node> node, const std::
                        {"KERNEL_3x3", dai::MedianFilter::KERNEL_3x3},
                        {"KERNEL_5x5", dai::MedianFilter::KERNEL_5x5},
                        {"KERNEL_7x7", dai::MedianFilter::KERNEL_7x7}};
+    presetModeMap = {{"TOF_LOW_RANGE", dai::ImageFiltersPresetMode::TOF_LOW_RANGE},
+                    {"TOF_MID_RANGE", dai::ImageFiltersPresetMode::TOF_MID_RANGE},
+                    {"TOF_HIGH_RANGE", dai::ImageFiltersPresetMode::TOF_HIGH_RANGE}};
 }
 ToFParamHandler::~ToFParamHandler() = default;
-void ToFParamHandler::declareParams(std::shared_ptr<dai::node::ToF> tof) {
+void ToFParamHandler::declareParams(std::shared_ptr<dai::node::ToF> tof, dai::CameraBoardSocket socket) {
     declareAndLogParam<bool>("i_publish_topic", true);
     declareAndLogParam<bool>("i_synced", false);
     declareAndLogParam<bool>("i_low_bandwidth", false);
@@ -32,36 +35,14 @@ void ToFParamHandler::declareParams(std::shared_ptr<dai::node::ToF> tof) {
     declareAndLogParam<bool>("i_reverse_stereo_socket_order", false);
     declareAndLogParam<std::string>("i_calibration_file", "");
     declareAndLogParam<int>("i_max_q_size", 8);
+    declareAndLogParam<int>("i_width", 640);
+    declareAndLogParam<int>("i_height", 400);
+    auto fps = declareAndLogParam<float>("i_fps", 15.0);
 
-    int socket = declareAndLogParam<int>("i_board_socket_id", static_cast<int>(dai::CameraBoardSocket::CAM_A));
-    // cam->setBoardSocket(static_cast<dai::CameraBoardSocket>(socket));
-    // cam->setSize(declareAndLogParam<int>("i_width", 640), declareAndLogParam<int>("i_height", 480));
-    // cam->setFps(declareAndLogParam<int>("i_fps", 30));
-    // auto tofConf = tof->initialConfig.get();
-    // if(declareAndLogParam<bool>("i_enable_optical_correction", false)) {
-    //     tofConf.enableOpticalCorrection = true;
-    // }
-    // if(declareAndLogParam<bool>("i_enable_fppn_correction", false)) {
-    //     tofConf.enableFPPNCorrection = true;
-    // }
-    // if(declareAndLogParam<bool>("i_enable_temperature_correction", false)) {
-    //     tofConf.enableTemperatureCorrection = true;
-    // }
-    // if(declareAndLogParam<bool>("i_enable_wiggle_correction", false)) {
-    //     tofConf.enableWiggleCorrection = true;
-    // }
-    // if(declareAndLogParam<bool>("i_enable_phase_unwrapping", false)) {
-    //     tofConf.enablePhaseUnwrapping = true;
-    // }
-    //
-    // tofConf.enablePhaseShuffleTemporalFilter = declareAndLogParam<bool>("i_enable_phase_shuffle_temporal_filter", true);
-    // tofConf.phaseUnwrappingLevel = declareAndLogParam<int>("i_phase_unwrapping_level", 4);
-    // tofConf.phaseUnwrapErrorThreshold = declareAndLogParam<int>("i_phase_unwrap_error_threshold", 100);
-    // std::vector<dai::MedianFilter> medianSettings = {
-    //     dai::MedianFilter::MEDIAN_OFF, dai::MedianFilter::KERNEL_3x3, dai::MedianFilter::KERNEL_5x5, dai::MedianFilter::KERNEL_7x7};
-    // tofConf.median = utils::getValFromMap(declareAndLogParam<std::string>("i_median_filter", "MEDIAN_OFF"), medianFilterMap);
-    //
-    // tof->initialConfig.set(tofConf);
+    auto sock = static_cast<dai::CameraBoardSocket>(declareAndLogParam<int>("i_board_socket_id", static_cast<int>(socket)));
+    dai::ImageFiltersPresetMode presetMode = utils::getValFromMap(declareAndLogParam<std::string>("i_preset_mode", "TOF_MID_RANGE"), presetModeMap);
+    tof->build(sock, presetMode, fps);
+
 }
 
 }  // namespace param_handlers
